Implementation of Salvage Logging
Added new stand-ranking method and two new requirements. The
stand-ranking method goes by time since last disturbance - wind, fire,
or both and if the average time since the last disturbance across a
stand's sites are greater than or equal to the time specified by the
user, then the stand is set aside for harvesting. This is an
implementation of salvage logging, issue #1 in base-harvest should be
resolved now.