Mark Sites for Single Repeat
Included a data structure in each stand to track which cells were acted on in the first iteration of a single repeat prescription. These sites are tracked and referred to upon the repeat to guarantee that they are the same sites to be harvested again.

Added a flag to the prescription class to let its objects know if it is a single repeat prescription and whether or not it is in the repeat step.