Patch Cutting Fixes and New Parameter
Tweaked the patch cutting algorithm to allow harvests as long as one patch has been selected, rather than requiring the target percentage of area to be reached.

Also changed the code to create a border of sites around each patch that will not be cut.

Added a new parameter "ALLOWOVERLAP" which will allow a user to override the previous change and have patches be able to share borders with each other.